feat: access cheacker for event module

+++ src/event_module/router.py
@@ -21,11 +21,13 @@ from src.event_module.schemas import (
     TagCreate,
     TagUpdate,
 )
+from src.event_module.utils import check_university_event
 from src.schemas import Response
 from src.tour_module.database.tour_event.tour_event_responses import (
     TourEventResponseHandler,
 )
 from src.university_module.router import university_event_response_handler
+from src.utils import Role, access_denied, role_access
 
 event_router = APIRouter(prefix="/event", tags=["event"])
 category_router = APIRouter(prefix="/event/category", tags=["category"])
@@ -73,23 +75,54 @@ async def get_event_by_id(
 
 @event_router.post("/", response_model=Response)
 async def create_event(
-    event: EventCreate, session: AsyncSession = Depends(get_async_session)
+    user_role: Role,
+    event: EventCreate,
+    session: AsyncSession = Depends(get_async_session),
 ) -> Response:
-    return await event_response_handler.create(model_create=event, session=session)
+    if role_access[user_role] >= role_access[Role.UNIVERSITY]:
+        return await event_response_handler.create(model_create=event, session=session)
+    else:
+        return access_denied()
 
 
 @event_router.put("/{event_id}", response_model=Response)
 async def update_event(
-    event: EventUpdate, session: AsyncSession = Depends(get_async_session)
+    user_role: Role,
+    user_id: int | None,
+    event: EventUpdate,
+    session: AsyncSession = Depends(get_async_session),
 ) -> Response:
-    return await event_response_handler.update(model_update=event, session=session)
+    if role_access[user_role] == role_access[Role.ADMIN]:
+        return await event_response_handler.update(model_update=event, session=session)
+    elif role_access[user_role] == role_access[Role.UNIVERSITY] and user_id is not None:
+        if await check_university_event(
+            university_id=user_id, event_id=event.id, session=session
+        ):
+            return await event_response_handler.update(
+                model_update=event, session=session
+            )
+
+    return access_denied()
 
 
 @event_router.delete("/{event_id}", response_model=Response)
 async def delete_event(
-    event_id: int, session: AsyncSession = Depends(get_async_session)
+    user_role: Role,
+    user_id: int | None,
+    event_id: int,
+    session: AsyncSession = Depends(get_async_session),
 ) -> Response:
-    return await event_response_handler.delete(model_id=event_id, session=session)
+    if role_access[user_role] == role_access[Role.ADMIN]:
+        return await event_response_handler.delete(model_id=event_id, session=session)
+    elif role_access[user_role] == role_access[Role.UNIVERSITY] and user_id is not None:
+        if await check_university_event(
+            university_id=user_id, event_id=event_id, session=session
+        ):
+            return await event_response_handler.delete(
+                model_id=event_id, session=session
+            )
+
+    return access_denied()
 
 
 @category_router.get("/", response_model=Response)
@@ -110,27 +143,44 @@ async def get_category_by_id(
 
 @category_router.post("/", response_model=Response)
 async def create_category(
-    category: CategoryCreate, session: AsyncSession = Depends(get_async_session)
+    user_role: Role,
+    category: CategoryCreate,
+    session: AsyncSession = Depends(get_async_session),
 ) -> Response:
-    return await category_response_handler.create(
-        model_create=category, session=session
-    )
+    if role_access[user_role] >= role_access[Role.ADMIN]:
+        return await category_response_handler.create(
+            model_create=category, session=session
+        )
+    else:
+        return access_denied()
 
 
 @category_router.put("/{category_id}", response_model=Response)
 async def update_category(
-    category: CategoryUpdate, session: AsyncSession = Depends(get_async_session)
+    user_role,
+    category: CategoryUpdate,
+    session: AsyncSession = Depends(get_async_session),
 ) -> Response:
-    return await category_response_handler.update(
-        model_update=category, session=session
-    )
+    if role_access[user_role] >= role_access[Role.ADMIN]:
+        return await category_response_handler.update(
+            model_update=category, session=session
+        )
+    else:
+        return access_denied()
 
 
 @category_router.delete("/{category_id}", response_model=Response)
 async def delete_category(
-    category_id: int, session: AsyncSession = Depends(get_async_session)
+    user_role: Role,
+    category_id: int,
+    session: AsyncSession = Depends(get_async_session),
 ) -> Response:
-    return await category_response_handler.delete(model_id=category_id, session=session)
+    if role_access[user_role] >= role_access[Role.ADMIN]:
+        return await category_response_handler.delete(
+            model_id=category_id, session=session
+        )
+    else:
+        return access_denied()
 
 
 @tag_router.get("/", response_model=Response)
@@ -149,32 +199,54 @@ async def get_tag_by_id(
 
 @tag_router.post("/", response_model=Response)
 async def create_tag(
-    tag: TagCreate, session: AsyncSession = Depends(get_async_session)
+    user_role: Role, tag: TagCreate, session: AsyncSession = Depends(get_async_session)
 ) -> Response:
-    return await tag_response_handler.create(model_create=tag, session=session)
+    if role_access[user_role] >= role_access[Role.ADMIN]:
+        return await tag_response_handler.create(model_create=tag, session=session)
+    else:
+        return access_denied()
 
 
 @tag_router.put("/{tag_id}", response_model=Response)
 async def update_tag(
-    tag: TagUpdate, session: AsyncSession = Depends(get_async_session)
+    user_role: Role, tag: TagUpdate, session: AsyncSession = Depends(get_async_session)
 ) -> Response:
-    return await tag_response_handler.update(model_update=tag, session=session)
+    if role_access[user_role] >= role_access[Role.ADMIN]:
+        return await tag_response_handler.update(model_update=tag, session=session)
+    else:
+        return access_denied()
 
 
 @tag_router.delete("/{tag_id}", response_model=Response)
 async def delete_tag(
-    tag_id: int, session: AsyncSession = Depends(get_async_session)
+    user_role: Role, tag_id: int, session: AsyncSession = Depends(get_async_session)
 ) -> Response:
-    return await tag_response_handler.delete(model_id=tag_id, session=session)
+    if role_access[user_role] >= role_access[Role.ADMIN]:
+        return await tag_response_handler.delete(model_id=tag_id, session=session)
+    else:
+        return access_denied()
 
 
-@event_router.post("/{event_id}/set_tags", response_model=Response)
+@event_router.post("/{event_id}/tags", response_model=Response)
 async def set_tags(
-    event_tag: EventTagListCreate, session: AsyncSession = Depends(get_async_session)
+    user_role: Role,
+    user_id: int | None,
+    event_tag: EventTagListCreate,
+    session: AsyncSession = Depends(get_async_session),
 ) -> Response:
-    return await event_tag_response_handler.create_list(
-        model_create=event_tag, session=session
-    )
+    if role_access[user_role] == role_access[Role.ADMIN]:
+        return await event_tag_response_handler.create_list(
+            model_create=event_tag, session=session
+        )
+    elif role_access[user_role] == role_access[Role.UNIVERSITY] and user_id is not None:
+        if await check_university_event(
+            university_id=user_id, event_id=event_tag.event_id, session=session
+        ):
+            return await event_tag_response_handler.create_list(
+                model_create=event_tag, session=session
+            )
+
+    return access_denied()
 
 
 @event_router.get("/tour_filter/{tour_id}", response_model=Response)
@@ -206,11 +278,24 @@ async def get_event_id_list_by_tag_id(
 
 @event_router.delete("/{event_id}/event_tag", response_model=Response)
 async def delete_event_tag(
-    event_tags: EventTagListDelete, session: AsyncSession = Depends(get_async_session)
+    user_role: Role,
+    user_id: int | None,
+    event_tags: EventTagListDelete,
+    session: AsyncSession = Depends(get_async_session),
 ) -> Response:
-    return await event_tag_response_handler.delete_by_delete_schema(
-        model_delete=event_tags, session=session
-    )
+    if role_access[user_role] == role_access[Role.ADMIN]:
+        return await event_tag_response_handler.delete_by_delete_schema(
+            model_delete=event_tags, session=session
+        )
+    elif role_access[user_role] == role_access[Role.UNIVERSITY] and user_id is not None:
+        if await check_university_event(
+            university_id=user_id, event_id=event_tags.event_id, session=session
+        ):
+            return await event_tag_response_handler.delete_by_delete_schema(
+                model_delete=event_tags, session=session
+            )
+
+    return access_denied()
 
 
 @event_router.get("/university_filter/{university_id}", response_model=Response)
